Simon Game
In this project I used HTML, CSS, and JavaScript (jQuery).
This is version 1.0 where the application can store the randomly generated object color through a javascript function, store it in an array and compare it with the options selected by the user and also store the user's answers in an array. Whenever the options generated by the code and the options selected by the user match, a level in the game increases and the function generates one more color adding it to the array. When the user fails the sequence of answers, a screen will be displayed indicating that the answer is wrong, and it will give the option to select any key on the keyboard to restart the game.